Moved Icelandic parsing to separate module
